Resolves several warnings that occur when attempting to build for 64-bit arch.

+++ desiccant/Classes/Zest/Foundation/NSIndexPath+Zest.m
@@ -11,8 +11,8 @@
 @implementation NSIndexPath (Zest)
 
 - (NSIndexPath *)relativeIndexPathWithBaseSection:(NSUInteger)baseSection andBaseRow:(NSUInteger)baseRow {
-	NSAssert2 (self.section >= baseSection, @"Requested relative index path given base section %d, which is greater than absolute section %d", baseSection, self.section);
-	NSAssert3 (self.section != baseSection || self.row >= baseRow, @"Requested relative index path given a base section equal to the absolute section %d, and a base row %d greater than the absolute section %d", baseSection, baseRow, self.row);
+	NSAssert2 (self.section >= baseSection, @"Requested relative index path given base section %lu, which is greater than absolute section %lu", (unsigned long) baseSection, (unsigned long) self.section);
+	NSAssert3 (self.section != baseSection || self.row >= baseRow, @"Requested relative index path given a base section equal to the absolute section %lu, and a base row %lu greater than the absolute section %lu", (unsigned long) baseSection, (unsigned long) baseRow, (unsigned long) self.row);
 	NSUInteger newSection = self.section - baseSection;
 	NSUInteger newRow;
 	if (self.section == baseSection) newRow = self.row - baseRow;
